Move ITreeDataSource impl. to TreeStoreBase
